updates paths,  adds back in eslint, moves components
