+++ apps/client-portal/_components/card-details.tsx
