+++ apps/copilot-crm/.eslintrc.js
@@ -6,4 +6,18 @@ module.exports = {
   parserOptions: {
     project: true,
   },
+  rules: {
+    "react/no-unescaped-entities": "off",
+    "@typescript-eslint/no-empty-interface": "off", // Changed from no-empty-object-type
+    "@typescript-eslint/no-explicit-any": "warn", // Changed to warn instead of off
+    "@typescript-eslint/no-unused-vars": ["warn", { argsIgnorePattern: "^_" }],
+    "prettier/prettier": ["error", {}, { usePrettierrc: true }],
+    "react/react-in-jsx-scope": "off",
+    "react/prop-types": "off",
+    // Add these rules for better TypeScript and React practices
+    "@typescript-eslint/explicit-function-return-type": "warn",
+    "@typescript-eslint/no-non-null-assertion": "warn",
+    "react-hooks/rules-of-hooks": "error",
+    "react-hooks/exhaustive-deps": "warn",
+  },
 };
