+++ apps/copilot-crm/_components/automation-info-card.tsx
