+++ apps/copilot-crm/_components/sidenav.tsx
