+++ packages/mocks/package.json
@@ -5,5 +5,8 @@
   "license": "MIT",
   "publishConfig": {
     "access": "public"
+  },
+  "dependencies": {
+    "msw": "^2.4.9"
   }
 }
