+++ packages/ui/package.json
@@ -37,6 +37,7 @@
     "@radix-ui/react-toggle-group": "^1.1.0",
     "@tanstack/react-table": "^8.20.5",
     "class-variance-authority": "^0.7.0",
+    "react-hook-form": "^7.53.0",
     "clsx": "^2.1.1",
     "cmdk": "1.0.0",
     "date-fns": "^4.1.0",
